Avance

+++ app/src/main/java/dev/neil/proyecto_final/Nav/ui/home/Adapter.kt
@@ -6,5 +6,5 @@ data class Adapter (
     val NombrePaseo:String,
     val DesripcionPaseo:String,
     val Tiempo:String,
-    val Rate:Int
+    val Rate: Float
 )
